UI... 70%

+++ src/main/java/com/example/demo/domain/Speciality.java
@@ -20,7 +20,7 @@ public class Speciality extends Identified {
     private Short placesQuantity;
 
     @Column
-    private BigDecimal certificateWeight;
+    private Double certificateWeight;
 
     @ManyToOne
     @JoinColumn(name = "facultyId", nullable = false)
